chore: move modal component location
