+++ packages/slice-machine/src/features/customTypes/customTypesBuilder/GenerateSliceWithAiModal/SliceCard.tsx
@@ -1,4 +1,5 @@
-import { Card, CardFooter, CardMedia } from "../Card";
+import { Card, CardFooter, CardMedia } from "@/components/Card";
+
 import { Slice } from "./GenerateSliceWithAiModal";
 
 interface SliceCardProps {
