+++ packages/slice-machine/src/features/customTypes/customTypesBuilder/GenerateSliceWithAiModal/index.tsx
